html/browsers/history/the-location-interface/location-prototype-setting-cross-origin.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>[[SetPrototypeOf]] on a Location object should not allow changing its value: cross-origin Location</title>
<link rel="help" href="http://html.spec.whatwg.org/multipage/#location-setprototypeof">
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<iframe src="//{{domains[www]}}:{{ports[http][1]}}/common/blank.html"></iframe>

<script>
"use strict";
setup({ explicit_done: true });

window.onload = () => {
  const targetLocation = frames[0].location;

  test(() => {
    assert_equals(Object.getPrototypeOf(targetLocation), null);
  }, "The prototype is null");

  test(() => {
    assert_throws(new TypeError, () => {
      Object.setPrototypeOf(targetLocation, {});
    });
  }, "Setting the prototype via Object.setPrototypeOf should throw a TypeError");

  test(() => {
    assert_throws(new TypeError, function() {
      targetLocation.__proto__ = {};
    });
  }, "Setting the prototype via __proto__ should throw a TypeError");

  test(() => {
    assert_false(Reflect.setPrototypeOf(targetLocation, {}));
  }, "Setting the prototype via Reflect.setPrototypeOf should return false");

  test(() => {
    assert_equals(Object.getPrototypeOf(targetLocation), null);
  }, "The prototype must still be null");

  test(() => {
    Object.setPrototypeOf(targetLocation, null);
  }, "Setting the prototype to null via Object.setPrototypeOf should not throw");

  test(() => {
    targetLocation.__proto__ = null;
  }, "Setting the prototype to null via __proto__ should not throw");

  test(() => {
    assert_true(Reflect.setPrototypeOf(targetLocation, null));
  }, "Setting the prototype to null via Reflect.setPrototypeOf should return true");

  done();
};
</script>
